chore: general clean up
