+++ packages/components/src/elements/core/readme.md
@@ -4,11 +4,11 @@
 
 ## Properties
 
-| Property        | Attribute | Description                                                 | Type                                                                                                                                                                                                           | Default     |
-| --------------- | --------- | ----------------------------------------------------------- | -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | ----------- |
-| `configuration` | --        | Configuration information like authentication configuration | `RVConfiguration`                                                                                                                                                                                              | `undefined` |
-| `filter`        | --        | Filter information to query the backend APIs                | `RVBaseFilterBusinessDateFrequencyType \| RVBaseFilterBusinessDateType \| RVBaseFilterBusinessType \| RVBaseFilterConnectionDateFrequencyType \| RVBaseFilterConnectionDateType \| RVBaseFilterConnectionType` | `undefined` |
-| `options`       | --        | For whitelabeling styling                                   | `RVOptions`                                                                                                                                                                                                    | `undefined` |
+| Property        | Attribute | Description                                                 | Type                                                                                                                                                                                                                                                                             | Default     |
+| --------------- | --------- | ----------------------------------------------------------- | -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | ----------- |
+| `configuration` | --        | Configuration information like authentication configuration | `RVConfiguration`                                                                                                                                                                                                                                                                | `undefined` |
+| `filter`        | --        | Filter information to query the backend APIs                | `RVFilterAccountingMethod \| RVFilterBalanceSheet \| RVFilterBankAccount \| RVFilterBills \| RVFilterCashflowStatements \| RVFilterCreditScore \| RVFilterDate \| RVFilterExpenses \| RVFilterFinancialRatio \| RVFilterIncomeStatements \| RVFilterInvoices \| RVFilterRevenue` | `undefined` |
+| `options`       | --        | For whitelabeling styling                                   | `RVOptions`                                                                                                                                                                                                                                                                      | `undefined` |
 
 ## Dependencies
