+++ packages/components/src/elements/gauge-chart/readme.md
@@ -4,11 +4,11 @@
 
 ## Properties
 
-| Property                     | Attribute | Description                                                 | Type                                                             | Default     |
-| ---------------------------- | --------- | ----------------------------------------------------------- | ---------------------------------------------------------------- | ----------- |
-| `configuration` _(required)_ | --        | Configuration information like authentication configuration | `RVConfiguration`                                                | `undefined` |
-| `filter` _(required)_        | --        | Filter information to query the backend APIs                | `RVBaseFilterBusinessDateType \| RVBaseFilterConnectionDateType` | `undefined` |
-| `options`                    | --        | For whitelabeling styling                                   | `RVOptions`                                                      | `undefined` |
+| Property                     | Attribute | Description                                                 | Type                 | Default     |
+| ---------------------------- | --------- | ----------------------------------------------------------- | -------------------- | ----------- |
+| `configuration` _(required)_ | --        | Configuration information like authentication configuration | `RVConfiguration`    | `undefined` |
+| `filter` _(required)_        | --        | Filter information to query the backend APIs                | `RVFilterRailzScore` | `undefined` |
+| `options`                    | --        | For whitelabeling styling                                   | `RVOptions`          | `undefined` |
 
 ## Dependencies
