+++ packages/components/src/helpers/test/chart.utils.spec.ts
@@ -1,20 +1,19 @@
-import {
-  RVAccountingProviders,
-  RVDateFilters,
-  RVFilter,
-  RVOptions,
-  RVReportFrequency,
-  RVReportTypes,
-} from '../../types';
+import { RVAccountingProviders, RVOptions, RVReportFrequency, RVReportTypes } from '../../types';
 import {
   getConfiguration,
   getDateFilter,
-  parseFilter,
   getOptions,
   validateBusinessParams,
   validateReportFrequencyParams,
   validateRequiredParams,
+  getFilter,
 } from '../chart.utils';
+
+import {
+  RVFilterBalanceSheet,
+  RVFilterBankAccount,
+  RVFilterAll,
+} from '../../types/interface/filter';
 
 describe('Chart Utils Helper', () => {
   describe('getConfiguration', () => {
@@ -68,43 +67,43 @@ describe('Chart Utils Helper', () => {
     });
   });
 
-  describe('parseFilter', () => {
+  describe('getFilter', () => {
     describe('using non object', () => {
       describe('failure path', () => {
         test('returns undefined due to integer passed in string', async () => {
-          expect(parseFilter('1')).toBeUndefined();
+          expect(getFilter('1')).toBeUndefined();
         });
 
         test('returns undefined due to plain string passed', async () => {
-          expect(parseFilter('testing param')).toBeUndefined();
+          expect(getFilter('testing param')).toBeUndefined();
         });
 
         test('returns undefined due to empty stringified array with passed', async () => {
-          expect(parseFilter('[]')).toBeUndefined();
+          expect(getFilter('[]')).toBeUndefined();
         });
 
         test('returns undefined due to stringified array passed', async () => {
-          expect(parseFilter('["data"]')).toBeUndefined();
+          expect(getFilter('["data"]')).toBeUndefined();
         });
 
         test('returns undefined due to empty stringified object passed', async () => {
-          expect(parseFilter('{}')).toBeUndefined();
+          expect(getFilter('{}')).toBeUndefined();
         });
 
         test('returns undefined due to stringified object passed without filter', async () => {
-          expect(parseFilter('{"tom": "works"}')).toBeUndefined();
+          expect(getFilter('{"tom": "works"}')).toBeUndefined();
         });
       });
 
       describe('success path', () => {
         test('returns formatted data due to the presence of filter', async () => {
-          const filter: RVFilter = {
+          const filter: RVFilterBankAccount = {
             reportType: RVReportTypes.BANK_ACCOUNT,
             businessName: 'businessName',
             serviceName: RVAccountingProviders.QUICKBOOKS,
           };
 
-          expect(parseFilter(filter)).toStrictEqual({
+          expect(getFilter(filter as RVFilterAll)).toStrictEqual({
             businessName: 'businessName',
             reportType: 'bankAccounts',
             serviceName: 'quickbooks',
@@ -115,25 +114,25 @@ describe('Chart Utils Helper', () => {
     describe('using object', () => {
       describe('failure path', () => {
         test('returns undefined due to stringified object passed without businessName', async () => {
-          const filter: RVFilter = {
+          const filter: RVFilterBankAccount = {
             reportType: RVReportTypes.BANK_ACCOUNT,
             businessName: '',
             serviceName: RVAccountingProviders.QUICKBOOKS,
           };
 
-          expect(parseFilter(JSON.stringify(filter))).toBeUndefined();
+          expect(getFilter(JSON.stringify(filter))).toBeUndefined();
         });
       });
 
       describe('success path', () => {
         test('returns formatted data due to the presence of filter', async () => {
-          const filter: RVFilter = {
+          const filter: RVFilterBankAccount = {
             reportType: RVReportTypes.BANK_ACCOUNT,
             businessName: 'businessName',
             serviceName: RVAccountingProviders.QUICKBOOKS,
           };
 
-          expect(parseFilter(JSON.stringify(filter))).toStrictEqual({
+          expect(getFilter(JSON.stringify(filter))).toStrictEqual({
             businessName: 'businessName',
             reportType: 'bankAccounts',
             serviceName: 'quickbooks',
@@ -146,28 +145,29 @@ describe('Chart Utils Helper', () => {
   describe('validateRequiredParams', () => {
     describe('success path', () => {
       test('returns true to valid params (businessName/serviceName)', async () => {
-        const filter: RVFilter = {
+        const filter: RVFilterBankAccount = {
           reportType: RVReportTypes.BANK_ACCOUNT,
           businessName: 'businessName',
           serviceName: RVAccountingProviders.QUICKBOOKS,
         };
-        expect(validateRequiredParams(filter)).toEqual(true);
+        expect(validateRequiredParams(filter as RVFilterAll)).toEqual(true);
       });
 
-      test('returns true to valid params (connectionId)', async () => {
-        const filter: RVFilter = {
-          reportType: RVReportTypes.BANK_ACCOUNT,
-          connectionId: 'connectionId',
-        };
-        expect(validateRequiredParams(filter)).toEqual(true);
-      });
+      // TODO: When connection is ready on web-backend
+      // test('returns true to valid params (connectionId)', async () => {
+      //   const filter: RVFilterBankAccount = {
+      //     reportType: RVReportTypes.BANK_ACCOUNT,
+      //     // connectionId: 'connectionId',
+      //   };
+      //   expect(validateRequiredParams(filter)).toEqual(true);
+      // });
     });
   });
 
   describe('validateReportFrequencyParams', () => {
     describe('success path', () => {
       test('returns true for correct params', async () => {
-        const filter: RVFilter = {
+        const filter: RVFilterBalanceSheet = {
           reportType: RVReportTypes.BALANCE_SHEET,
           businessName: '',
           serviceName: RVAccountingProviders.QUICKBOOKS,
@@ -175,7 +175,7 @@ describe('Chart Utils Helper', () => {
           startDate: '2020-01-01',
           endDate: '2020-01-01',
         };
-        expect(validateReportFrequencyParams(filter)).toEqual(true);
+        expect(validateReportFrequencyParams(filter as RVFilterAll)).toEqual(true);
       });
     });
   });
@@ -183,56 +183,58 @@ describe('Chart Utils Helper', () => {
   describe('validateBusinessParams', () => {
     describe('failure path', () => {
       test('returns false if there is not businessName', async () => {
-        const filter: RVFilter = {
+        const filter: RVFilterBankAccount = {
           reportType: RVReportTypes.BANK_ACCOUNT,
           businessName: '',
           serviceName: RVAccountingProviders.QUICKBOOKS,
         };
-        expect(validateBusinessParams(filter)).toEqual(false);
+        expect(validateBusinessParams(filter as RVFilterAll)).toEqual(false);
       });
       test('returns false if there is not businessName/serviceName', async () => {
-        const filter: RVFilter = {
+        const filter: RVFilterBankAccount = {
           reportType: RVReportTypes.BANK_ACCOUNT,
           businessName: '',
           serviceName: undefined,
         };
-        expect(validateBusinessParams(filter)).toEqual(false);
+        expect(validateBusinessParams(filter as RVFilterAll)).toEqual(false);
       });
-      test('returns false if there is not businessName', async () => {
-        const filter: RVFilter = {
-          reportType: RVReportTypes.BANK_ACCOUNT,
-          connectionId: '',
-        };
-        expect(validateBusinessParams(filter)).toEqual(false);
-      });
+      // TODO: When connection is ready on web-backend
+      // test('returns false if there is not businessName', async () => {
+      //   const filter: RVFilterBankAccount = {
+      //     reportType: RVReportTypes.BANK_ACCOUNT,
+      //     connectionId: '',
+      //   };
+      //   expect(validateBusinessParams(filter)).toEqual(false);
+      // });
     });
 
     describe('success path', () => {
       test('returns true for businessName and serviceName', async () => {
-        const filter: RVFilter = {
+        const filter: RVFilterBankAccount = {
           reportType: RVReportTypes.BANK_ACCOUNT,
           businessName: 'businessNames',
           serviceName: RVAccountingProviders.QUICKBOOKS,
         };
-        expect(validateBusinessParams(filter)).toEqual(true);
+        expect(validateBusinessParams(filter as RVFilterAll)).toEqual(true);
       });
 
       test('returns false for businessName with not serviceName', async () => {
-        const filter: RVFilter = {
+        const filter: RVFilterBankAccount = {
           reportType: RVReportTypes.BANK_ACCOUNT,
           businessName: 'businessNames',
           serviceName: undefined,
         };
-        expect(validateBusinessParams(filter)).toEqual(false);
+        expect(validateBusinessParams(filter as RVFilterAll)).toEqual(false);
       });
 
-      test('returns true for connectionId', async () => {
-        const filter: RVFilter = {
-          reportType: RVReportTypes.BANK_ACCOUNT,
-          connectionId: 'connectionId',
-        };
-        expect(validateBusinessParams(filter)).toEqual(true);
-      });
+      // TODO: When connection is ready on web-backend
+      // test('returns true for connectionId', async () => {
+      //   const filter: RVFilterBankAccount = {
+      //     reportType: RVReportTypes.BANK_ACCOUNT,
+      //     connectionId: 'connectionId',
+      //   };
+      //   expect(validateBusinessParams(filter)).toEqual(true);
+      // });
     });
   });
 
@@ -298,7 +300,7 @@ describe('Chart Utils Helper', () => {
         });
 
         test('returns undefined for the same date', async () => {
-          const filter: RVDateFilters = {
+          const filter: RVFilterBalanceSheet = {
             reportType: RVReportTypes.BALANCE_SHEET,
             businessName: 'businessName',
             serviceName: RVAccountingProviders.QUICKBOOKS,
@@ -310,7 +312,7 @@ describe('Chart Utils Helper', () => {
         });
 
         test('returns undefined for startDate after than endDate', async () => {
-          const filter: RVDateFilters = {
+          const filter: RVFilterBalanceSheet = {
             reportType: RVReportTypes.BALANCE_SHEET,
             businessName: 'businessName',
             serviceName: RVAccountingProviders.QUICKBOOKS,
@@ -324,7 +326,7 @@ describe('Chart Utils Helper', () => {
 
       describe('success path', () => {
         test('returns formatted data due to the presence of filter', async () => {
-          const filter: RVDateFilters = {
+          const filter: RVFilterBalanceSheet = {
             reportType: RVReportTypes.BALANCE_SHEET,
             businessName: 'businessName',
             serviceName: RVAccountingProviders.QUICKBOOKS,
@@ -351,7 +353,7 @@ describe('Chart Utils Helper', () => {
           //   businessName: '',
           //   serviceName: RVAccountingProviders.QUICKBOOKS,
           // };
-          // expect(parseFilter(JSON.stringify(filter))).toBeUndefined();
+          // expect(getFilter(JSON.stringify(filter))).toBeUndefined();
         });
       });
 
@@ -362,7 +364,7 @@ describe('Chart Utils Helper', () => {
           //   businessName: 'businessName',
           //   serviceName: RVAccountingProviders.QUICKBOOKS,
           // };
-          // expect(parseFilter(JSON.stringify(filter))).toStrictEqual({
+          // expect(getFilter(JSON.stringify(filter))).toStrictEqual({
           //   businessName: 'businessName',
           //   reportType: 'bankAccounts',
           //   serviceName: 'quickbooks',
