+++ packages/components/src/types/interface/filter.ts
@@ -1,128 +1,171 @@
-// import { RVFrequencyTypes, RVServiceProviders } from '../enum';
-// import { RVReportFrequency } from '../enum/date';
-// import { RVNoDateTypes, RVNoFrequencyTypes } from '../enum/report-type';
+import { RVServiceProviders, RVReportFrequency, RVReportTypes, RVAccountingMethod } from '../enum';
 
-import { RVReportTypes } from '../enum';
+export interface RVFilterBusiness {
+  /**
+   * businessName: Unique business name identifier.
+   * required **if connectionId is not passed**
+   */
+  businessName: string;
+  /**
+   * serviceName: Name of accounting platform service.
+   * required **if connectionId is not passed**
+   * (sandbox, quickBooks, quickBooksDesktop, freshBooks,
+   * xero, oracleNetsuite, sageBusinessCloud, sageIntacct, dynamicsBusinessCentral)
+   */
+  serviceName: RVServiceProviders;
+}
 
-// export interface RVBaseFilterBusiness {
-//   /**
-//    * businessName: Unique business name identifier.
-//    * required **if connectionId is not passed**
-//    */
-//   businessName: string;
-//   /**
-//    * serviceName: Name of accounting platform service.
-//    * required **if connectionId is not passed**
-//    * (sandbox, quickBooks, quickBooksDesktop, freshBooks,
-//    * xero, oracleNetsuite, sageBusinessCloud, sageIntacct, dynamicsBusinessCentral)
-//    */
-//   serviceName: RVServiceProviders;
-// }
+/**
+ * RVFilterConnection
+ * Not currently being used, since web-backend hasn't implemented this
+ */
+export interface RVFilterConnection {
+  /**
+   * connectionId: Unique Connection identifier.
+   * required **if businessName and serviceName are not passed**
+   */
+  connectionId: string;
+}
 
-// export interface RVBaseFilterConnection {
-//   /**
-//    * connectionId: Unique Connection identifier.
-//    * required **if businessName and serviceName are not passed**
-//    */
-//   connectionId: string;
-// }
+export interface RVFilterDate {
+  /**
+   * startDate: Date from which the data begins
+   * format: **yyyy-MM-dd**
+   */
+  startDate: string; // required based on report type,
+  /**
+   * endDate: Date which the data ends
+   * format: **yyyy-MM-dd**
+   */
+  endDate: string; // required based on report type,
+}
 
-// export interface RVBaseFilterBusinessDate extends RVBaseFilterBusiness {
-//   /**
-//    * startDate: Date from which the data begins
-//    * format: **yyyy-MM-dd**
-//    */
-//   startDate: string; // required based on report type,
-//   /**
-//    * endDate: Date which the data ends
-//    * format: **yyyy-MM-dd**
-//    */
-//   endDate: string; // required based on report type,
-// }
+export interface RVFilterReportFrequency {
+  /**
+   * reportFrequency: The snapshot frequency (month, quarter, year)
+   */
+  reportFrequency: RVReportFrequency;
+}
 
-// export interface RVBaseFilterConnectionDate extends RVBaseFilterConnection {
-//   /**
-//    * startDate: Date from which the data begins
-//    * format: **yyyy-MM-dd**
-//    */
-//   startDate: string;
-//   /**
-//    * endDate: Date which the data ends
-//    * format: **yyyy-MM-dd**
-//    */
-//   endDate: string;
-// }
+export interface RVFilterAccountingMethod {
+  /**
+   * accountingMethod: the enum for accounting methods: accrual or cash
+   */
+  accountingMethod: RVAccountingMethod;
+}
 
-// export interface RVBaseFilterConnectionDateFrequency extends RVBaseFilterConnectionDate {
-//   /**
-//    * reportFrequency: The snapshot frequency (month, quarter, year)
-//    */
-//   reportFrequency: RVReportFrequency;
-// }
+export interface RVFilterReportType {
+  /**
+   * reportType: the enum for to say the type of the report to be displayed
+   */
+  reportType: RVReportTypes;
+}
 
-// export interface RVBaseFilterBusinessDateFrequency extends RVBaseFilterBusinessDate {
-//   /**
-//    * reportFrequency: The snapshot frequency (month, quarter, year)
-//    */
-//   reportFrequency: RVReportFrequency;
-// }
+export interface RVFilterBalanceSheet
+  extends RVFilterReportType,
+    RVFilterBusiness,
+    RVFilterDate,
+    RVFilterReportFrequency {
+  reportType: RVReportTypes.BALANCE_SHEET;
+}
 
-// export interface RVBaseFilterConnectionType extends RVBaseFilterBusiness {
-//   /**
-//    * reportType: Type of report, bankAccounts, creditScores
-//    */
-//   reportType: RVNoDateTypes;
-// }
+export interface RVFilterBankAccount extends RVFilterReportType, RVFilterBusiness {
+  reportType: RVReportTypes.BANK_ACCOUNT;
+}
 
-// export interface RVBaseFilterBusinessType extends RVBaseFilterConnection {
-//   /**
-//    * reportType: Type of report, bankAccounts, creditScores
-//    */
-//   reportType: RVNoDateTypes;
-// }
+export interface RVFilterBills extends RVFilterReportType, RVFilterBusiness, RVFilterDate {
+  reportType: RVReportTypes.BILLS;
+}
 
-// export interface RVBaseFilterConnectionDateFrequencyType
-//   extends RVBaseFilterConnectionDateFrequency {
-//   /**
-//    * reportType: Type of report, cashflowStatements, balanceSheets, incomeStatements
-//    */
-//   reportType: RVFrequencyTypes;
-// }
+export interface RVFilterCashflowStatements
+  extends RVFilterReportType,
+    RVFilterBusiness,
+    RVFilterDate,
+    RVFilterReportFrequency {
+  reportType: RVReportTypes.CASHFLOW_STATEMENTS;
+}
 
-// export interface RVBaseFilterBusinessDateFrequencyType extends RVBaseFilterBusinessDateFrequency {
-//   /**
-//    * reportType: Type of report, cashflowStatements, balanceSheets, incomeStatements
-//    */
-//   reportType: RVFrequencyTypes;
-// }
+export interface RVFilterCreditScore extends RVFilterReportType, RVFilterBusiness, RVFilterDate {
+  reportType: RVReportTypes.CREDIT_SCORE;
+}
 
-// export interface RVBaseFilterBusinessDateType extends RVBaseFilterBusinessDate {
-//   /**
-//    * reportType: Type of transaction, invoices, bills
-//    */
-//   reportType: RVNoFrequencyTypes;
-// }
+export interface RVFilterExpenses
+  extends RVFilterReportType,
+    RVFilterBusiness,
+    RVFilterDate,
+    RVFilterReportFrequency {
+  reportType: RVReportTypes.EXPENSES;
+}
 
-// export interface RVBaseFilterConnectionDateType extends RVBaseFilterConnectionDate {
-//   /**
-//    * reportType: Type of transaction, invoices, bills
-//    */
-//   reportType: RVNoFrequencyTypes;
-// }
+export interface RVFilterFinancialRatio
+  extends RVFilterReportType,
+    RVFilterBusiness,
+    RVFilterDate,
+    RVFilterReportFrequency {
+  reportType: RVReportTypes.FINANCIAL_RATIO;
+}
 
-// export type RVBaseFilter = RVBaseFilterBusiness | RVBaseFilterConnection;
-// export type RVBaseFilterDate = RVBaseFilterConnectionDate | RVBaseFilterBusinessDate;
-// export type RVBaseFilterDateFrequency =
-//   | RVBaseFilterConnectionDateFrequency
-//   | RVBaseFilterBusinessDateFrequency;
-// export type RVBaseAllFilter = RVBaseFilter | RVBaseFilterDate | RVBaseFilterDateFrequency;
-// export type RVFilterFrequency =
-//   | RVBaseFilterConnectionDateFrequencyType
-//   | RVBaseFilterBusinessDateFrequencyType;
-// export type RVFilterDate = RVBaseFilterConnectionDateType | RVBaseFilterBusinessDateType;
-// export type RVFilterType = RVBaseFilterConnectionType | RVBaseFilterBusinessType;
-// export type RVFilter = RVFilterFrequency | RVFilterDate | RVFilterType;
-// export type RVDateFilters = RVFilterDate | RVFilterFrequency;
-// export type RVAllFilter = RVFilterDate | RVFilterFrequency;
+export interface RVFilterIncomeStatements
+  extends RVFilterReportType,
+    RVFilterBusiness,
+    RVFilterDate,
+    RVFilterReportFrequency {
+  reportType: RVReportTypes.INCOME_STATEMENTS;
+}
 
-export type abc = RVReportTypes.BALANCE_SHEET;
+export interface RVFilterInvoices
+  extends RVFilterReportType,
+    RVFilterBusiness,
+    RVFilterDate,
+    RVFilterReportFrequency {
+  reportType: RVReportTypes.INVOICES;
+}
+
+export interface RVFilterRevenue
+  extends RVFilterReportType,
+    RVFilterBusiness,
+    RVFilterDate,
+    RVFilterReportFrequency {
+  reportType: RVReportTypes.REVENUE;
+}
+
+export interface RVFilterRailzScore
+  extends RVFilterReportType,
+    RVFilterBusiness,
+    RVFilterDate,
+    RVFilterReportFrequency {
+  reportType: RVReportTypes.RAILZ_SCORE;
+}
+
+export interface RVFilterAll
+  extends RVFilterBusiness,
+    RVFilterConnection,
+    RVFilterDate,
+    RVFilterReportFrequency,
+    RVFilterAccountingMethod,
+    RVFilterReportType {}
+
+export type RVFilterAllReportTypes =
+  | RVFilterBalanceSheet
+  | RVFilterBankAccount
+  | RVFilterBills
+  | RVFilterCashflowStatements
+  | RVFilterCreditScore
+  | RVFilterExpenses
+  | RVFilterFinancialRatio
+  | RVFilterFinancialRatio
+  | RVFilterIncomeStatements
+  | RVFilterInvoices
+  | RVFilterRevenue
+  | RVFilterDate
+  | RVFilterDate
+  | RVFilterDate
+  | RVFilterAccountingMethod;
+
+export type RVFilterTransactions = RVFilterInvoices | RVFilterBills;
+export type RVFilterStatements =
+  | RVFilterBalanceSheet
+  | RVFilterIncomeStatements
+  | RVFilterCashflowStatements;
+export type RVFilterGauge = RVFilterRailzScore;
+export type RVFilterPie = RVFilterExpenses | RVFilterRevenue;
