+++ packages/components/src/types/interface/index.ts
@@ -1,5 +1,5 @@
 export * from './configuration';
 export * from './filter';
-export * from './filter2';
+export * from './filter';
 export * from './option';
 export * from './summary';
